LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/scroll-timeline-on-pseudo-like-part-expected.txt



FAIL timeline.source does not expose pseudo-like part assert_equals: expected "chooser" but got ""
